Fixed application freezing after password deletion confirmation
Go channels are unbuffered by default. This means that goroutine sending message to unbuffered channel is blocking until some receiver reads from given channel. Problem was that goroutine was waiting to send message but receiver did not read as because it is in a event loop and no event were taking place - the chicken and egg problem. Adding buffer for a message unblocked writing goroutine, giving possibility for program to get to the receiver and consume message.